sort updated
Bug was found in the sorting section.  Previously,  it only sorted the A column (to eliminate blank cells and group like symbols).  However the dates were not sorted.  This affected the percentage change because the first date of the array for the opening price may not be the earliest date.   

now it's been coded to sort column A then sort column B.